+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 
+## [v0.3] - 2022-04-25
+### Added
+- Waiting for headers
+- Move operations for `Request` and `Response`
+- Synchronize `Session` with `boost::asio::strand`
+
+### Changed
+- Function `async_wait()` renamed to `async_await_completion()`
+- Split logic into `Request` and `Response` classes
+
 ## [v0.2]
 ### Added
 - JSON writer and reader
@@ -14,5 +24,6 @@
 ### Fixed
 - Cookie share support
 
-[Unreleased]: https://github.com/terrakuh/curlio/compare/v0.2..dev
+[Unreleased]: https://github.com/terrakuh/curlio/compare/v0.3..dev
+[v0.3]: https://github.com/terrakuh/curlio/compare/v0.2..v0.3
 [v0.2]: https://github.com/terrakuh/curlio/compare/v0.1..v0.2